update (CGFloat) explicit casts to .f casts

+++ components/ActionSheet/src/Theming/MDCActionSheetController+MaterialTheming.m
@@ -15,9 +15,9 @@
 #import "MDCActionSheetController+MaterialTheming.h"
 #import "MaterialColor.h"
 
-static const CGFloat kHighAlpha = (CGFloat)0.87;
-static const CGFloat kMediumAlpha = (CGFloat)0.6;
-static const CGFloat kInkAlpha = (CGFloat)0.16;
+static const CGFloat kHighAlpha = 0.87f;
+static const CGFloat kMediumAlpha = 0.6f;
+static const CGFloat kInkAlpha = 0.16f;
 
 @implementation MDCActionSheetController (MaterialTheming)
