+++ components/AppBar/tests/unit/Theming/MDCAppBarViewController_MaterialThemingTests.m
@@ -39,16 +39,16 @@
   self.containerScheme = [[MDCContainerScheme alloc] init];
   MDCSemanticColorScheme *colorScheme =
       [[MDCSemanticColorScheme alloc] initWithDefaults:MDCColorSchemeDefaultsMaterial201804];
-  colorScheme.primaryColor = [UIColor colorWithWhite:(CGFloat)0.9 alpha:0];
-  colorScheme.primaryColorVariant = [UIColor colorWithWhite:(CGFloat)0.8 alpha:(CGFloat)0.1];
-  colorScheme.secondaryColor = [UIColor colorWithWhite:(CGFloat)0.7 alpha:(CGFloat)0.2];
-  colorScheme.errorColor = [UIColor colorWithWhite:(CGFloat)0.6 alpha:(CGFloat)0.3];
-  colorScheme.surfaceColor = [UIColor colorWithWhite:(CGFloat)0.5 alpha:(CGFloat)0.4];
-  colorScheme.backgroundColor = [UIColor colorWithWhite:(CGFloat)0.4 alpha:(CGFloat)0.5];
-  colorScheme.onPrimaryColor = [UIColor colorWithWhite:(CGFloat)0.3 alpha:(CGFloat)0.6];
-  colorScheme.onSecondaryColor = [UIColor colorWithWhite:(CGFloat)0.2 alpha:(CGFloat)0.7];
-  colorScheme.onSurfaceColor = [UIColor colorWithWhite:(CGFloat)0.1 alpha:(CGFloat)0.8];
-  colorScheme.onBackgroundColor = [UIColor colorWithWhite:0 alpha:(CGFloat)0.9];
+  colorScheme.primaryColor = [UIColor colorWithWhite:0.9f alpha:0];
+  colorScheme.primaryColorVariant = [UIColor colorWithWhite:0.8f alpha:0.1f];
+  colorScheme.secondaryColor = [UIColor colorWithWhite:0.7f alpha:0.2f];
+  colorScheme.errorColor = [UIColor colorWithWhite:0.6f alpha:0.3f];
+  colorScheme.surfaceColor = [UIColor colorWithWhite:0.5f alpha:0.4f];
+  colorScheme.backgroundColor = [UIColor colorWithWhite:0.4f alpha:0.5f];
+  colorScheme.onPrimaryColor = [UIColor colorWithWhite:0.3f alpha:0.6f];
+  colorScheme.onSecondaryColor = [UIColor colorWithWhite:0.2f alpha:0.7f];
+  colorScheme.onSurfaceColor = [UIColor colorWithWhite:0.1f alpha:0.8f];
+  colorScheme.onBackgroundColor = [UIColor colorWithWhite:0 alpha:0.9f];
   self.containerScheme.colorScheme = colorScheme;
 
   MDCTypographyScheme *typographyScheme =
@@ -109,14 +109,14 @@
                         self.containerScheme.typographyScheme.headline6);
   XCTAssertEqualObjects(
       self.appBarController.navigationBar.titleTextColor,
-      [self.containerScheme.colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.87]);
+      [self.containerScheme.colorScheme.onSurfaceColor colorWithAlphaComponent:0.87f]);
   XCTAssertEqualObjects(self.appBarController.headerView.backgroundColor,
                         self.containerScheme.colorScheme.surfaceColor);
   XCTAssertEqualObjects(self.appBarController.navigationBar.leadingBarItemsTintColor,
                         self.containerScheme.colorScheme.onSurfaceColor);
   XCTAssertEqualObjects(
       self.appBarController.navigationBar.trailingBarItemsTintColor,
-      [self.containerScheme.colorScheme.onSurfaceColor colorWithAlphaComponent:(CGFloat)0.54]);
+      [self.containerScheme.colorScheme.onSurfaceColor colorWithAlphaComponent:0.54f]);
   XCTAssertNil(self.appBarController.headerView.shadowLayer);
 }
